+++ src/extra/libs/sjpg/tjpgd.c
@@ -26,6 +26,7 @@
 
 #include "tjpgd.h"
 #if LV_USE_SJPG
+#include "tjdec.h"
 
 #if JD_FASTDECODE == 2
 #define HUFF_BIT	10	/* Bit length to apply fast huffman decode */
